+++ packages/orq-rc/src/orq_ai_sdk/models/listagentsop.py
@@ -261,7 +261,7 @@ class ListAgentsKnowledgeBases(BaseModel):
     configuration: ListAgentsKnowledgeBaseConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K6JFC6M7BBTDMVV5H8XPRX5D"
+    id: Optional[str] = "01K6JK9V1ECE9ZM1QAXW824DDZ"
     r"""The id of the resource"""